upgrade to react-scripts 4.0.1
I'm naively hoping that this fixes the issue with Emotion/Chakra styles that only shows up on prod…

We've had an issue with prod-only bugs before, and it was due to a compiler bug re comments in JSX. I'm wondering if maybe one of our deps now has the same or a similar bug? and if maybe upgrading react-scripts will upgrade webpack, which might fix it?

It's a bit of a longshot, but it's a good upgrade to do anyway, so let's see if we get lucky!

Also, react-scripts 4 has Fast Refresh built in, so we can remove our overrides that added it!